Basic Events Model-Repository-Controller Ready

+++ hub/src/main/java/lol/fmg/hub/repositories/eventsRepositories/EventCategoryRepository.java
@@ -3,5 +3,8 @@ package lol.fmg.hub.repositories.eventsRepositories;
 import lol.fmg.hub.models.events.EventCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventCategoryRepository extends JpaRepository<EventCategory, Integer> {
+import java.util.List;
+
+public interface EventCategoryRepository extends JpaRepository<EventCategory, Long> {
+    // Custom queries can be added here if required
 }
