+++ hub/src/main/java/lol/fmg/hub/repositories/eventsRepositories/EventTagRepository.java
@@ -3,5 +3,8 @@ package lol.fmg.hub.repositories.eventsRepositories;
 import lol.fmg.hub.models.events.EventTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventTagRepository extends JpaRepository<EventTag, Integer> {
+import java.util.List;
+
+public interface EventTagRepository extends JpaRepository<EventTag, Long> {
+    // Custom queries can be added here if required
 }
